requirements.txt
python==3
Django==4.1.1
sqlparse==0.4.3
djangorestframework
twilio
python-decouple

